Hydrator ignores private/protected getter/setter
If the entity had a getter/setter which was private or protected the hydrator tried to call it anyway
and thus threw an exception. We now silently ignore them if there are not callable.

This allow for entities that have "read-only" fields which are set internally (such as timestamp).

This is a backport of https://github.com/doctrine/DoctrineModule/pull/560